chat_bot_api/requirements.txt
fastapi~=0.115.9
uvicorn
python-dotenv~=1.1.0
requests~=2.32.3
psycopg2-binary~=2.9.10
pandas~=2.2.3
nltk~=3.9.1
sentence-transformers~=4.1.0
rapidfuzz~=3.13.0
langdetect~=1.0.9

psycopg2~=2.9.10
numpy~=2.2.5
scikit-learn~=1.6.1
bs4~=0.0.2
beautifulsoup4~=4.13.4
deep-translator~=1.11.4
pydantic~=2.11.4
translator~=0.0.9
qdrant-client~=1.14.2
tqdm~=4.67.1
chromadb~=1.0.9
langchain~=0.3.25